refactor: extract indexer/scan.rs (phase 3, step 10)
Move workspace scanning logic from indexer.rs to src/indexer/scan.rs:
- IndexingGuard RAII struct (clears indexing_in_progress on drop)
- DEFAULT_MAX_INDEX_FILES const and resolve_max_files()
- find_files_for_types() helper (rg-based type→file lookup)
- Indexer::index_workspace / index_workspace_full / index_workspace_prioritized
- Indexer::index_workspace_impl (core: discover → cache partition → concurrent parse → LSP progress)
- Indexer::save_cache_to_disk

scan_tests.rs (8 new tests):
- resolve_max_files env var presence/absence/invalid cases (with ENV_LOCK)
- find_files_for_types empty-input fast-path
- find_files_for_types locates class in real .kt file (rg integration)
- IndexingGuard drop clears indexing_in_progress flag
- index_workspace_full populates definitions map
- concurrent second run is aborted when indexing_in_progress is set

cache_tests.rs:
- workspace_cache_path_stable and workspace_cache_path_differs now use
  with_xdg_cache to hold XDG_CACHE_LOCK; fixes a race with other tests
  that set XDG_CACHE_HOME concurrently
